+++ cmake/FindZstd.cmake
@@ -50,3 +50,15 @@ if(ZSTD_FOUND AND NOT TARGET zstd::zstd)
       INTERFACE_INCLUDE_DIRECTORIES ${ZSTD_INCLUDE_DIR}
   )
 endif()
+
+# Check if the Conan-provided target exists
+if(TARGET zstd::libzstd_static)
+  # If the Conan target exists, use it
+  set(ZSTD_TARGET zstd::libzstd_static)
+elseif(TARGET zstd::zstd)
+  # If the system-installed target exists, use it
+  set(ZSTD_TARGET zstd::zstd)
+else()
+  # Error out if neither target is found
+  message(FATAL_ERROR "Zstd target not found.")
+endif()